refactor(runtime): extract turn/protocol_detect.rs into agent/turn — #7415
